requirements.txt
DateTime==5.4
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Session==0.8.0
google-ai-generativelanguage==0.6.4
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
